backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyMySQL==1.1.0
Pillow==12.0.0
django-filter==23.3
python-dotenv==1.0.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.2
markdown==3.4.4